온도 관련 변수 타입 Float -> Double로 수정

+++ src/main/java/com/google/smarthome/controller/AppServerController.java
@@ -84,8 +84,8 @@ public class AppServerController {
                 result.getDeviceId(), Map.of(
                         "currentModeSettings", Map.of("mode_boiler", result.getModeValue()),
                         "online", true,
-                        "temperatureAmbientCelsius", Float.parseFloat(result.getCurrentTemp()),
-                        "temperatureSetpointCelsius", Float.parseFloat(result.getTempStatus()),
+                        "temperatureAmbientCelsius", Double.parseDouble(result.getCurrentTemp()),
+                        "temperatureSetpointCelsius", Double.parseDouble(result.getTempStatus()),
                         "on", powerOnOff));
 
         log.info("Constructed device state: " + deviceStates);
